requirements.txt
python-telegram-bot==13.11
python-dotenv==0.20.0
google-cloud-dialogflow==2.13.0
 